+++ src/dipdup/database.py
@@ -194,44 +194,56 @@ async def generate_schema(
     conn: SupportedClient,
     name: str,
 ) -> None:
-    if isinstance(conn, AsyncpgClient):
-        await pg_create_schema(conn, name)
+    if isinstance(conn, SqliteClient):
+        await Tortoise.generate_schemas()
+    elif isinstance(conn, AsyncpgClient):
+        await _pg_create_schema(conn, name)
+        await Tortoise.generate_schemas()
+        await _pg_create_functions(conn)
+        await _pg_create_views(conn)
+    else:
+        raise NotImplementedError
 
-    await Tortoise.generate_schemas()
 
-    if isinstance(conn, AsyncpgClient):
-        # NOTE: Create a view for monitoring head status
-        sql_path = Path(__file__).parent / 'sql' / 'dipdup_head_status.sql'
-        # TODO: Configurable interval
-        await execute_sql(conn, sql_path, HEAD_STATUS_TIMEOUT)
+async def _pg_create_functions(conn: AsyncpgClient) -> None:
+    for fn in (
+        'dipdup_approve.sql',
+        'dipdup_wipe.sql',
+        # TODO: Alias, remove in 8.0
+        'truncate_schema.sql',
+    ):
+        sql_path = Path(__file__).parent / 'sql' / fn
+        await execute_sql(conn, sql_path)
 
 
-async def _wipe_schema_postgres(
+async def _pg_create_views(conn: AsyncpgClient) -> None:
+    sql_path = Path(__file__).parent / 'sql' / 'dipdup_head_status.sql'
+    # TODO: Configurable interval
+    await execute_sql(conn, sql_path, HEAD_STATUS_TIMEOUT)
+
+
+async def _pg_wipe_schema(
     conn: AsyncpgClient,
     schema_name: str,
     immune_tables: set[str],
 ) -> None:
     immune_schema_name = f'{schema_name}_immune'
 
-    # NOTE: Create a truncate_schema function to trigger cascade deletion
-    sql_path = Path(__file__).parent / 'sql' / 'truncate_schema.sql'
-    await execute_sql(conn, sql_path, schema_name, immune_schema_name)
-
     # NOTE: Move immune tables to a separate schema - it's free!
     if immune_tables:
-        await pg_create_schema(conn, immune_schema_name)
+        await _pg_create_schema(conn, immune_schema_name)
         for table in immune_tables:
-            await pg_move_table(conn, table, schema_name, immune_schema_name)
+            await _pg_move_table(conn, table, schema_name, immune_schema_name)
 
     await conn.execute_script(f"SELECT truncate_schema('{schema_name}')")
 
     if immune_tables:
         for table in immune_tables:
-            await pg_move_table(conn, table, immune_schema_name, schema_name)
-        await pg_drop_schema(conn, immune_schema_name)
+            await _pg_move_table(conn, table, immune_schema_name, schema_name)
+        await _pg_drop_schema(conn, immune_schema_name)
 
 
-async def _wipe_schema_sqlite(
+async def _sqlite_wipe_schema(
     conn: SqliteClient,
     path: str,
     immune_tables: set[str],
@@ -271,23 +283,23 @@ async def wipe_schema(
     """Truncate schema preserving immune tables. Executes in a transaction"""
     async with conn._in_transaction() as conn:
         if isinstance(conn, SqliteClient):
-            await _wipe_schema_sqlite(conn, schema_name, immune_tables)
+            await _sqlite_wipe_schema(conn, schema_name, immune_tables)
         elif isinstance(conn, AsyncpgClient):
-            await _wipe_schema_postgres(conn, schema_name, immune_tables)
+            await _pg_wipe_schema(conn, schema_name, immune_tables)
         else:
             raise NotImplementedError
 
 
-async def pg_create_schema(conn: AsyncpgClient, name: str) -> None:
+async def _pg_create_schema(conn: AsyncpgClient, name: str) -> None:
     """Create PostgreSQL schema if not exists"""
     await conn.execute_script(f'CREATE SCHEMA IF NOT EXISTS {name}')
 
 
-async def pg_drop_schema(conn: AsyncpgClient, name: str) -> None:
+async def _pg_drop_schema(conn: AsyncpgClient, name: str) -> None:
     await conn.execute_script(f'DROP SCHEMA IF EXISTS {name}')
 
 
-async def pg_move_table(conn: AsyncpgClient, name: str, schema: str, new_schema: str) -> None:
+async def _pg_move_table(conn: AsyncpgClient, name: str, schema: str, new_schema: str) -> None:
     """Move table from one schema to another"""
     await conn.execute_script(f'ALTER TABLE {schema}.{name} SET SCHEMA {new_schema}')